test(refactor): move fixtures to conftest
