+++ tests/unit/test_routes_api_status.py
@@ -1,188 +1,4 @@
-from unittest.mock import Mock
-
 import pytest
-from celery.result import AsyncResult, GroupResult
-
-from sketch_map_tool.exceptions import MapGenerationError, QRCodeError
-from sketch_map_tool.routes import app
-
-
-@pytest.fixture()
-def client():
-    return app.test_client()
-
-
-@pytest.fixture()
-def mock_request_task_mapping(uuid, monkeypatch):
-    """Mock request id to task id mapping."""
-    monkeypatch.setattr(
-        "sketch_map_tool.routes.db_client_flask.get_async_result_id",
-        lambda *_: uuid,
-    )
-
-
-@pytest.fixture()
-def mock_async_result_success(monkeypatch):
-    mock = Mock(spec=AsyncResult)
-    mock.status = "SUCCESS"
-    mock.ready.return_value = True
-    mock.failed.return_value = False
-    mock.successful.return_value = True
-
-    monkeypatch.setattr("sketch_map_tool.routes.celery_app.AsyncResult", lambda _: mock)
-    return mock
-
-
-@pytest.fixture()
-def mock_async_result_started(monkeypatch):
-    mock = Mock(spec=AsyncResult)
-    mock.status = "STARTED"
-    mock.ready.return_value = False
-    mock.failed.return_value = False
-    mock.successful.return_value = False
-    monkeypatch.setattr("sketch_map_tool.routes.celery_app.AsyncResult", lambda _: mock)
-    return mock
-
-
-@pytest.fixture
-def mock_async_result_failure(monkeypatch):
-    """Mock task result wich failed w/ expected error"""
-    mock = Mock(spec=AsyncResult)
-    mock.status = "FAILURE"
-    mock.ready.return_value = True
-    mock.failed.return_value = True
-    mock.successful.return_value = False
-    mock.get.side_effect = MapGenerationError("Mock error")
-    monkeypatch.setattr("sketch_map_tool.routes.celery_app.AsyncResult", lambda _: mock)
-    return mock
-
-
-@pytest.fixture
-def mock_async_result_failure_upload_processing(monkeypatch):
-    """Mock task result wich failed w/ expected error"""
-    mock = Mock(spec=AsyncResult)
-    mock.status = "FAILURE"
-    mock.ready.return_value = True
-    mock.failed.return_value = True
-    mock.successful.return_value = False
-    mock.get.side_effect = QRCodeError("Mock error")
-
-    monkeypatch.setattr("sketch_map_tool.routes.celery_app.AsyncResult", lambda _: mock)
-    return mock
-
-
-@pytest.fixture
-def mock_async_result_failure_hard(monkeypatch):
-    """Mock task result wich failed w/ unexpected error"""
-    mock = Mock(spec=AsyncResult)
-    mock.status = "FAILURE"
-    mock.ready.return_value = True
-    mock.failed.return_value = True
-    mock.successful.return_value = False
-    mock.get.side_effect = ValueError()
-
-    monkeypatch.setattr("sketch_map_tool.routes.celery_app.AsyncResult", lambda _: mock)
-    return mock
-
-
-@pytest.fixture
-def mock_group_result_success(mock_async_result_success, monkeypatch):
-    mock = Mock(spec=GroupResult)
-    mock.ready.return_value = True
-    mock.failed.return_value = False
-    mock.successful.return_value = True
-    mock.results = [mock_async_result_success]
-
-    monkeypatch.setattr(
-        "sketch_map_tool.routes.celery_app.GroupResult.restore", lambda _: mock
-    )
-    return mock
-
-
-@pytest.fixture
-def mock_group_result_started(mock_async_result_started, monkeypatch):
-    mock = Mock(spec=GroupResult)
-    mock.ready.return_value = False
-    mock.failed.return_value = False
-    mock.successful.return_value = True
-    mock.results = [mock_async_result_started]
-
-    monkeypatch.setattr(
-        "sketch_map_tool.routes.celery_app.GroupResult.restore", lambda _: mock
-    )
-    return mock
-
-
-@pytest.fixture
-def mock_group_result_failure(mock_async_result_failure_upload_processing, monkeypatch):
-    mock = Mock(spec=GroupResult)
-    mock.ready.return_value = True
-    mock.failed.return_value = True
-    mock.successful.return_value = False
-    mock.results = [mock_async_result_failure_upload_processing]
-    mock.get.side_effect = mock_async_result_failure_upload_processing.get
-
-    monkeypatch.setattr(
-        "sketch_map_tool.routes.celery_app.GroupResult.restore", lambda _: mock
-    )
-
-
-@pytest.fixture
-def mock_group_result_failure_hard(mock_async_result_failure_hard, monkeypatch):
-    mock = Mock(spec=GroupResult)
-    mock.ready.return_value = True
-    mock.failed.return_value = True
-    mock.successful.return_value = False
-    mock.results = [mock_async_result_failure_hard]
-    mock.get.side_effect = mock_async_result_failure_hard.get
-
-    monkeypatch.setattr(
-        "sketch_map_tool.routes.celery_app.GroupResult.restore", lambda _: mock
-    )
-
-
-@pytest.fixture
-def mock_group_result_started_success_failure(
-    mock_async_result_started,
-    mock_async_result_success,
-    mock_async_result_failure_upload_processing,
-    monkeypatch,
-):
-    mock = Mock(spec=GroupResult)
-    mock.ready.return_value = False
-    mock.failed.return_value = True
-    mock.successful.return_value = False
-    mock.results = [
-        mock_async_result_started,
-        mock_async_result_success,
-        mock_async_result_failure_upload_processing,
-    ]
-    mock.get.side_effect = mock_async_result_failure_upload_processing.get
-
-    monkeypatch.setattr(
-        "sketch_map_tool.routes.celery_app.GroupResult.restore", lambda _: mock
-    )
-
-
-@pytest.fixture
-def mock_group_result_success_failure(
-    mock_async_result_success,
-    mock_async_result_failure_upload_processing,
-    monkeypatch,
-):
-    mock = Mock(spec=GroupResult)
-    mock.ready.return_value = True
-    mock.failed.return_value = True
-    mock.successful.return_value = False
-    mock.results = [
-        mock_async_result_success,
-        mock_async_result_failure_upload_processing,
-    ]
-    mock.get.side_effect = mock_async_result_failure_upload_processing.get
-
-    monkeypatch.setattr(
-        "sketch_map_tool.routes.celery_app.GroupResult.restore", lambda _: mock
-    )
 
 
 @pytest.mark.parametrize("type_", ("sketch-map", "quality-report"))
@@ -234,7 +50,7 @@ def test_status_failure(
     assert resp.json["id"] == uuid
     assert resp.json["type"] == type_
     assert resp.json["status"] == "FAILURE"
-    assert resp.json["errors"] == ["MapGenerationError: Mock error"]
+    assert resp.json["errors"] == ["QRCodeError: Mock error"]
     assert "href" not in resp.json.keys()
 
 
@@ -298,6 +114,7 @@ def test_group_status_failure(
     assert resp.json["id"] == uuid
     assert resp.json["type"] == type_
     assert resp.json["status"] == "FAILURE"
+    assert resp.json["errors"] == ["QRCodeError: Mock error"]
 
 
 @pytest.mark.parametrize("type_", ("raster-results", "vector-results"))
